+++ sklearn/tests/test_common.py
@@ -1054,3 +1054,63 @@ def test_import_all_consistency():
                 raise AttributeError(
                     "Module '{}' has no attribute '{}'".format(
                         modname, name))
+
+
+def test_sparsify_estimators():
+    """Test if predict with sparsified estimators works.
+
+    Tests regression, binary classification, and multi-class classification.
+    """
+    estimators = all_estimators()
+    X = np.array([[-2, -1], [-1, -1], [-1, -2], [1, 1], [1, 2], [2, 1]])
+    y = [1, 1, 1, 2, 2, 2]
+
+    # test regression and binary classification
+    for name, Estimator in estimators:
+        try:
+            Estimator.sparsify
+        except:
+            continue
+
+        est = Estimator()
+
+        est.fit(X, y)
+        pred_orig = est.predict(X)
+
+        # test sparsify with dense inputs
+        est.sparsify()
+        assert_true(sparse.issparse(est.coef_))
+        pred = est.predict(X)
+        assert_array_equal(pred, pred_orig)
+
+        # pickle and unpickle with sparse coef_
+        est = pickle.loads(pickle.dumps(est))
+        assert_true(sparse.issparse(est.coef_))
+        pred = est.predict(X)
+        assert_array_equal(pred, pred_orig)
+
+
+    # test multiclass classification
+    classifiers = all_estimators(type_filter='classifier')
+    y[-1] = 3  # make multi-class
+    for name, Classifier in classifiers:
+        try:
+            Classifier.sparsify
+        except:
+            continue
+        est = Classifier()
+
+        est.fit(X, y)
+        pred_orig = est.predict(X)
+
+        # test sparsify with dense inputs
+        est.sparsify()
+        assert_true(sparse.issparse(est.coef_))
+        pred = est.predict(X)
+        assert_array_equal(pred, pred_orig)
+
+        # pickle and unpickle with sparse coef_
+        est = pickle.loads(pickle.dumps(est))
+        assert_true(sparse.issparse(est.coef_))
+        pred = est.predict(X)
+        assert_array_equal(pred, pred_orig)