+++ sklearn/utils/__init__.py
@@ -350,8 +350,11 @@ def gen_batches(n, batch_size):
         yield slice(start, n)
 
 
-def gen_even_slices(n, n_packs):
+def gen_even_slices(n, n_packs, n_samples=None):
     """Generator to create n_packs slices going up to n.
+
+    Pass n_samples when the slices are to be used for sparse matrix indexing;
+    slicing off-the-end raises an exception, while it works for NumPy arrays.
 
     Examples
     --------
@@ -372,6 +375,8 @@ def gen_even_slices(n, n_packs):
             this_n += 1
         if this_n > 0:
             end = start + this_n
+            if n_samples is not None:
+                end = min(n_samples, end)
             yield slice(start, end, None)
             start = end